LTO, binary index storage, other improvements

+++ main.h
@@ -1,6 +1,6 @@
 #ifndef PICO_EPAPER_73_MAIN_H
 #define PICO_EPAPER_73_MAIN_H
 
-unsigned int voltageToPercentage(float voltage);
+float voltageToPercentage(float voltage);
 
 #endif //PICO_EPAPER_73_MAIN_H
